Abstract the state configuration away from tree-only structure.
Motivation
----------
The active state configuration of a state machine can be either:
 - just one node for state machines only made of simple states
 - a list of nodes for state machines with composite but no orthogonal
   states.
 - a tree of nodes for states machines with orthogonal states.
The intent of this change is to abstract the state configuration such as
the optimal data structure and state machine state can be chosen on the
fly based on the complexity of the state machine.

Changes
-------
TreeStateMachineState was renamed and is now an abstract class.
The data structure is now an abstract StateConfiguration and its
traversal is left to the SimpleStateConfiguration or
OrthogonalStateConfiguration implementation.

Results
-------
Traversal of the active state configuration is abstracted away from the
state machine executor.