Prepare data structure for state machine export.
Motivation
----------
The state of the state machine needs to be fully exportable such as a
future resume function could reset the state of the state machine with
that export. Current StateMachineSnapshot is missing a few things, such
as backed up history and isn't easy to export due to its ties to
StateConfiguration.

Changes
-------
The new implementation of StateMachineSnapshot is now self-contained and
much closer to a POJO. It now contains a snapshot of history nodes. The
active state configuration of StateMachineSnapshot is no longer a
instance of StateConfiguration but is now a simple tree of strings,
each representing a state id.
All unit tests have been adjusted to use that new data structure and
ActiveState in the test package has been deleted.
The delay of some time transitions was increased from 10ms to 50ms to
limit failure on slower environments.

Results
-------
StateMachineSnapshot is now serializable and can be easily decorated (or
not) to export in json or other data format.
Some unit tests are more robust.