Test deep/shallow history for an orthogonal state.
Motivation
----------
Test coverage reveals the copy method of OrthogonalStateConfiguration is
never used. That is because this method is only invoked by the state
machine executor when it needs to save the history of a state. Looking
at the deep/shallow history tests, none indeed tests an orthogonal state.

Changes
-------
Two new unit tests to test deep and shallow history of an orthogonal
state. New method in StateMachineProcessorTester to test the history
snapshot. Some methods in StateMachineProcessorTester were renamed.

Results
-------
Increased test coverage. The two new unit tests passed with success the
first time and did not reveal any bugs.