Pause and resume state machine functionality.
Motivation
----------
Provide functionality to pause a state machine, save its entire state,
and resume execution from that point for the same, one different or many
other state machine executors of the same state machine.

Changes
-------
StateMachineExecutor now provides two new methods: pause and resume.
Implementation of those methods lead to adding an important missing
feature in PoolStateMachineExecutor, which is the ability to take a
snapshot at any point of time while avoiding data inconsistency.
Previous buggy implementation was allowing to take a snapshot in the
middle of a run-to-completion step.
The new pause and resume implementation has been tested with simple,
composite, orthogonal and submachine state machines as well as with
state machine which have either a shallow or deep history vertex.
The implemetnation of the test class ActiveStateTree was simplified and
all tests now take a snapshot to access the context of the state
machine.

Result
------
The state machine can be paused, forked by creating many new executors
of the same state machine, and all resumed from the same snapshot.
Importing and exporting a state machine state appears to be fully
functional.