Always return a copy of the context for snapshots.
Motivation
----------
Taking any snapshot must return a copy of the context when possible
otherwise this may yield getting non-sense data in the snapshot with an
active state configuration which is always a copy and a context which
keeps getting updated by the state machine executor.

Changes
-------
StateMachineSnapshotImpl makes a copy of the context when possible, i.e.
when the context implements Copyable. Some unit tests contexts got fixed
in response to that change.
More JavaDoc for a few interfaces in the behaviorstatemachines package.

Result
------
Calling snapshot or pause always returns a snapshot which holds a copy
of the context if that one implements Copyable.