Fix simple.Test11 and activity.Test11 unit tests
Motivation
----------
Althought passing on local machine, a couple of unit tests failed on
travis CI: simple.Test11 and activity.Test1. This is highly likely due
to the fact that those tests are time dependent.

Changes
-------
Moved the count down latch of those two tests to the effect of the last
transition of the state machine.

Result
------
Travis CI should be back to normal.