Abstract string logger and improve the quality of logs
Motivations
-----------
Implementing all listeners is a pain. This project contains a
StringBufferLogger in the test package that could be abstracted and
published. In addition, despite the numerous events AllListener can
catch, none help figuring out on what state the executor is and for
which executor.

Modifcations
------------
Abstracted StringBufferLogger and moved it to the sources. Added two new
listeners: StateEnterListener and StateExitListener which are invoked
even if the state does not have any enter/exit behaviors. New method in
StateMachineExecutor that returns the name of the executor.

Result
------
More and better information can be displayed when logging.