Implemented shutdown methods for StateMachineShardPoolExecutor
Motivations
-----------
StateMachineShardPoolExecutor is a kind of a custom thread pool but does
not provide shutdown methods at the moment.

Modifications
-------------
Changed how the Thread are built in StateMachineShardPoolExecutor.
Instead of using custom Workers extending Thread, it now creates
mono-threaded instances of ThreadPoolExecutor and passes to it a
LinkedBlockingQueue.

Result
------
The sharded thread pool executor can be easily shutdown.