Add a lock around scheduling and running activities
Motivations
-----------
Some unit tests often run into race conditions while trying to execute
activities on Travis CI. This results in a never ending (i.e forever
blocking) unit test.
The analysis revealed a bug in the way activities are scheduled and
resulting completion events triggered.
Scheduling activities requires performing two actions: 1) schedule the
activity in the thread pool; and 2) store the Future returned by the
thread pool in the state of the state machine. In some cases, the
activity can start and complete before the future is stored in the state
of the state machine which prevents the completion event at the end of
the activity to ever be fired.

Modifications
-------------
Added a lock around scheduling activities. The new lock belongs to
StateStatus.

Result
------
Race conditions around activities is now history.