Add a StdoutLogger
Motivations
-----------
Logging at the end of a test or run of a state machine isn't always
convenient, especially when the run fails/blocks somewhere before the
end of the execution.

Modifications
-------------
Implemented a new StdoutLogger that extends AbstractStringLogger. This
logger logs absolutely everything.
Renamed AbstractStringLoggerListener to AbstractStringLogger.
Renamed StringBufferLoggerListener to StringBufferLogger.

Result
------
Out of the box stdout logger for any state machines.