Initialize AbstractStateMachineExecutor in a thread-safe manner
Motivations
-----------
The StateMachineExecutor interface requires the implementer to define
setters to set the state machine, context and other vital attributes
AbstractStateMachineExecutor needs to run. This becomes a problem for
implementations of AbstractStateMachineExecutor that start and run
events in threads other than the creator-thread. In particular, the
PoolStateMachineExecutor failed to start multiple times because the
start method (in go) was processed quicker enough by the thread pool for
this one to not see the stateMachine attribute initialized.

Modifications
-------------
 - Builder pattern-only for AbstractStateMachineExecutor implementations
   and made private their constructors.
 - Removed setters from the StateMachineExecutor interface.
 - Made nearly all attributes of AbstractStateMachineExecutor final.
 - Removed the parameter of StateMachineExecutor#resume
 - New pause boolean and related methods in StateMachineState.
 - Disabled pause and resume for PoolStateMachineExecutor as the
   implementation is still unstable.

Result
------
Instances of AbstractStateMachineExecutor are now thread-safe when
created without snapshots.
Still not thread-safe are:
 - the AbstractStateMachineExecutor when created with a snapshot; the
   state is notstate is not thread-safely initialized yet.
 - ExecutorListener