Improved support of time events
Motivations
-----------
Time events have been limited to hardcoded constants for how long the
event has to wait before being activated. That is not always convenient
as some state machines might need to change the duration of that time
event based on the information in the context.

Modifications
-------------
Created a new FunctionRelativeTimeEvent event that takes a function of
the context and returns a Duration. A new overload of the 'after' method
is available to use it. Simplified the handling of time events in the
AbstractStateMachineExecutor such as all TimeEvents can be processed the
same way. Minor unit test coverage improvements.

Result
------
The duration of a time event might now be dynamically determined by the
running state machine as a function of its context.