Fixed race condition in StateMachineShardPoolExecutor
Motivations
-----------
Tests in StateMachineShardPoolExecutor often fail on Travis CI (e.g:
build 45, 47). The state machine either get stuck or throws an exception
saying go() wasn't invoked.

Modifications
-------------
StateMachineShardPoolExecutor was overriding and pooling the `start`
behavior. This was the root of this evil since it is the `go` method
that sets the `go` boolean to `true`. Therefore, the `go` method
would be invoked, `start` scheduled, `go` boolean set immediately to
true in the current thread and the change not visible in the thread
executing `start` and then `take`. The change consists in not pooling
the `start` method but instead pooling the `go` call. `go` is also used
in `resume` so in order to avoid scheduling in chain `resume` then `go`
whe nresuming a state machine, the behavior of the `go` method was moved
to an `internalGo` method so it can be directly invoked from
AbstractStateMachineExecutor::resume().

Results
-------
Race conditions noticed in builds 45 and 47 should no longer occur.